Fixed code style and made tests work.
