+++ lib/fs-cache.js
@@ -145,7 +145,7 @@ var cache = module.exports = function(params, callback) {
         function(result) {
           write(file, result, function(err) {
             callback(err, result);
-          })
+          });
         },
         callback
       );
